+++ package.json
@@ -746,6 +746,7 @@
     "@kbn/resizable-layout-examples-plugin": "link:examples/resizable_layout_examples",
     "@kbn/resolver-test-plugin": "link:x-pack/test/plugin_functional/plugins/resolver_test",
     "@kbn/response-ops-feature-flag-service": "link:packages/response-ops/feature_flag_service",
+    "@kbn/response-ops-rule-params": "link:packages/response-ops/rule_params",
     "@kbn/response-stream-plugin": "link:examples/response_stream",
     "@kbn/rison": "link:packages/kbn-rison",
     "@kbn/rollup": "link:x-pack/packages/rollup",
