+++ packages/response-ops/rule_params/README.md
@@ -1,0 +1,3 @@
+# @kbn/response-ops-rule-params
+
+The package is responsible for the parameters' schema of all rule types. The alerting plugin uses this package to generate OAS documentation for the `params` property in the rule in requests and responses.
