+++ packages/response-ops/rule_params/kibana.jsonc
@@ -1,0 +1,5 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/response-ops-rule-params",
+  "owner": "@elastic/response-ops"
+}
