+++ tsconfig.base.json
@@ -1448,6 +1448,8 @@
       "@kbn/resolver-test-plugin/*": ["x-pack/test/plugin_functional/plugins/resolver_test/*"],
       "@kbn/response-ops-feature-flag-service": ["packages/response-ops/feature_flag_service"],
       "@kbn/response-ops-feature-flag-service/*": ["packages/response-ops/feature_flag_service/*"],
+      "@kbn/response-ops-rule-params": ["packages/response-ops/rule_params"],
+      "@kbn/response-ops-rule-params/*": ["packages/response-ops/rule_params/*"],
       "@kbn/response-stream-plugin": ["examples/response_stream"],
       "@kbn/response-stream-plugin/*": ["examples/response_stream/*"],
       "@kbn/rison": ["packages/kbn-rison"],
