+++ x-pack/plugins/alerting/common/routes/backfill/response/schemas/v1.ts
@@ -6,6 +6,7 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { ruleParamsSchemaV1 } from '@kbn/response-ops-rule-params';
 import { adHocRunStatus } from '../../../../constants';
 
 export const statusSchema = schema.oneOf([
@@ -26,7 +27,7 @@ export const backfillResponseSchema = schema.object({
     name: schema.string(),
     tags: schema.arrayOf(schema.string()),
     rule_type_id: schema.string(),
-    params: schema.recordOf(schema.string(), schema.maybe(schema.any())),
+    params: ruleParamsSchemaV1,
     api_key_owner: schema.nullable(schema.string()),
     api_key_created_by_user: schema.maybe(schema.nullable(schema.boolean())),
     consumer: schema.string(),
