+++ x-pack/plugins/alerting/common/routes/rule/apis/create/schemas/v1.ts
@@ -6,6 +6,7 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { ruleParamsSchemaWithDefaultValueV1 } from '@kbn/response-ops-rule-params';
 import { validateDurationV1, validateHoursV1, validateTimezoneV1 } from '../../../validation';
 import { notifyWhenSchemaV1, alertDelaySchemaV1 } from '../../../response';
 import { alertsFilterQuerySchemaV1 } from '../../../../alerts_filter_query';
@@ -166,10 +167,7 @@ export const createBodySchema = schema.object({
       })
     )
   ),
-  params: schema.recordOf(schema.string(), schema.maybe(schema.any()), {
-    defaultValue: {},
-    meta: { description: 'The parameters for the rule.' },
-  }),
+  params: ruleParamsSchemaWithDefaultValueV1,
   schedule: schema.object(
     {
       interval: schema.string({
