+++ x-pack/plugins/alerting/common/routes/rule/response/types/v1.ts
@@ -6,22 +6,21 @@
  */
 
 import type { TypeOf } from '@kbn/config-schema';
+import { RuleParamsV1 } from '@kbn/response-ops-rule-params';
 import {
-  ruleParamsSchemaV1,
   ruleResponseSchemaV1,
   ruleSnoozeScheduleSchemaV1,
   ruleLastRunSchemaV1,
   monitoringSchemaV1,
 } from '..';
 
-export type RuleParams = TypeOf<typeof ruleParamsSchemaV1>;
 export type RuleSnoozeSchedule = TypeOf<typeof ruleSnoozeScheduleSchemaV1>;
 export type RuleLastRun = TypeOf<typeof ruleLastRunSchemaV1>;
 export type Monitoring = TypeOf<typeof monitoringSchemaV1>;
 
 type RuleResponseSchemaType = TypeOf<typeof ruleResponseSchemaV1>;
 
-export interface RuleResponse<Params extends RuleParams = never> {
+export interface RuleResponse<Params extends RuleParamsV1 = never> {
   id: RuleResponseSchemaType['id'];
   enabled: RuleResponseSchemaType['enabled'];
   name: RuleResponseSchemaType['name'];
