+++ x-pack/plugins/alerting/server/application/backfill/result/schemas/index.ts
@@ -6,6 +6,7 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { ruleParamsSchema } from '@kbn/response-ops-rule-params';
 import { adHocRunStatus } from '../../../../../common/constants';
 
 export const statusSchema = schema.oneOf([
@@ -32,7 +33,7 @@ export const backfillSchema = schema.object({
     name: schema.string(),
     tags: schema.arrayOf(schema.string()),
     alertTypeId: schema.string(),
-    params: schema.recordOf(schema.string(), schema.maybe(schema.any())),
+    params: ruleParamsSchema,
     apiKeyOwner: schema.nullable(schema.string()),
     apiKeyCreatedByUser: schema.maybe(schema.nullable(schema.boolean())),
     consumer: schema.string(),
