+++ x-pack/plugins/alerting/server/application/rule/methods/create/schemas/create_rule_data_schema.ts
@@ -6,6 +6,7 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { ruleParamsSchemaWithDefaultValue } from '@kbn/response-ops-rule-params';
 import { validateDuration } from '../../../validation';
 import {
   notifyWhenSchema,
@@ -23,7 +24,7 @@ export const createRuleDataSchema = schema.object(
     consumer: schema.string(),
     tags: schema.arrayOf(schema.string(), { defaultValue: [] }),
     throttle: schema.maybe(schema.nullable(schema.string({ validate: validateDuration }))),
-    params: schema.recordOf(schema.string(), schema.maybe(schema.any()), { defaultValue: {} }),
+    params: ruleParamsSchemaWithDefaultValue,
     schedule: schema.object({
       interval: schema.string({ validate: validateDuration }),
     }),
