+++ x-pack/plugins/alerting/server/application/rule/methods/update/schemas/update_rule_data_schema.ts
@@ -6,6 +6,7 @@
  */
 
 import { schema } from '@kbn/config-schema';
+import { ruleParamsSchemaWithDefaultValue } from '@kbn/response-ops-rule-params';
 import { validateDuration } from '../../../validation';
 import {
   notifyWhenSchema,
@@ -23,7 +24,7 @@ export const updateRuleDataSchema = schema.object(
       interval: schema.string({ validate: validateDuration }),
     }),
     throttle: schema.maybe(schema.nullable(schema.string({ validate: validateDuration }))),
-    params: schema.recordOf(schema.string(), schema.maybe(schema.any()), { defaultValue: {} }),
+    params: ruleParamsSchemaWithDefaultValue,
     actions: schema.arrayOf(actionRequestSchema, { defaultValue: [] }),
     systemActions: schema.maybe(schema.arrayOf(systemActionRequestSchema, { defaultValue: [] })),
     notifyWhen: schema.maybe(schema.nullable(notifyWhenSchema)),
